tidy default implementation
