--- examples/provider/event_stream.rs
@@ -1,24 +1,0 @@
-use anyhow::Result;
-use credibil_dwn::event::{Event, Subscriber};
-use credibil_dwn::provider::EventStream;
-use futures::stream::StreamExt;
-
-use super::ProviderImpl;
-
-const SUBJECT: &str = "events";
-
-impl EventStream for ProviderImpl {
-    /// Subscribe to a owner's event stream.
-    async fn subscribe(&self, owner: &str) -> Result<Subscriber> {
-        let subscriber = self.nats_client.subscribe(format!("{SUBJECT}.{owner}")).await?;
-        let mapped = subscriber.map(|m| serde_json::from_slice::<Event>(&m.payload).unwrap());
-        Ok(Subscriber::new(mapped))
-    }
-
-    /// Emits an event to a owner's event stream.
-    async fn emit(&self, owner: &str, event: &Event) -> Result<()> {
-        let bytes = serde_json::to_vec(event)?;
-        self.nats_client.publish(format!("{SUBJECT}.{owner}"), bytes.into()).await?;
-        Ok(())
-    }
-}
